Add alternatives, size and negative prompt support to spam mode
Anything in [] split by comma , will generate multiple generations, multiple [] groups will generate all variations of the two groups combined, this is recursive with any number of groups.
The size is set with {} like so {W:H} where the values will be normalized afterwards, the separator character can be anything except a period which is used for decimals.
The negative prompt will be the part left when splitting on semi-colon, ;
Added a prompt hint to the response message, if the spam mode is used with alternatives, the used alternatives are printed instead of a truncated prompt if the prompt does not fit.
This comes with a few new values in the config.